pyproject.toml
[tool.poetry]
name = "django-minutes"
version = "0.0.1"
description = "A django module providing a minutes taking app"
authors = ["Jonatan Zint <jonatan@ctrl.alt.coop>"]
license = "MIT"
packages = [
    { include = "minutes" },
]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.0"
djangorestframework = "^3.11.0"
django-ordered-model = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest-django = "^3.9.0"
ipdb = "^0.13.2"
tox = "^3.14.6"
pylint = "^2.4.4"
coverage = "^5.1"

[build-system]
requires = ["poetry>=0.12, <1"]
build-backend = "poetry.masonry.api"
